Made material section collapsible for certain pages - Not always needed to be visible
