+++ components/MaterialLists/MaterialList.tsx
@@ -1,182 +1,53 @@
-"use client";
-
-import groupMaterialsByMaterialType from "@/actions/material/group/groupMaterialsByMaterialType";
-import BlogsList from "@/components/MaterialLists/BlogsList";
-import CertificatesList from "@/components/MaterialLists/CertificatesList";
-import ModuleList from "@/components/MaterialLists/ModuleList";
-import ProjectsList from "@/components/MaterialLists/ProjectsList";
-import { Button } from "@/components/shadcn/ui/button";
 import {
-  BLOG_PAGE,
-  CERTIFICATES_PAGE,
-  EXPERIENCE_PAGE,
-  PROJECTS_PAGE,
-} from "@/constants/pages";
-import blogDatabase, { blogKeys } from "@/database/blogs";
-import certificateDatabase, { certificateKeys } from "@/database/certificates";
-import moduleDatabase, { moduleKeys } from "@/database/modules";
-import projectDatabase, { projectKeys } from "@/database/projects";
+  Accordion,
+  AccordionContent,
+  AccordionItem,
+  AccordionTrigger,
+} from "@/components/shadcn/ui/accordion";
 import MaterialType from "@/enums/MaterialType";
-import MaterialGroupInterface from "@/interfaces/material/MaterialGroupInterface";
-import MaterialInterface from "@/interfaces/material/MaterialInterface";
-import MaterialListProps from "@/interfaces/props/MaterialListProps";
-import Link from "next/link";
-import React, { useState } from "react";
-import {
-  Tabs,
-  TabsContent,
-  TabsList,
-  TabsTrigger,
-} from "@/components/shadcn/ui/tabs";
-import rolesDatabase, { roleKeys } from "@/database/roles";
-import WorkList from "./WorkList";
+import React from "react";
+import MaterialTab from "./MaterialTab";
 
-interface MaterialSectionInterface {
-  name: MaterialType;
-  materials: string[];
-  materialHashmap: Database<MaterialInterface>;
-  basePath?: string;
-  ListComponent: React.ComponentType<MaterialListProps>;
-}
-
-interface MaterialSectionProps {
+export interface MaterialTabsProps {
   materialKeys: string[];
   defaultTab?: MaterialType;
+  isCollapsible?: boolean;
+  sectionName?: string;
 }
 
-/**
- * Component displaying a list of all materials.
- * This includes:
- * - Projects
- * - Certificates
- * - Blogs
- * - Modules
- *
- * @param materialKeys The keys of the material to display.
- * @returns Section displaying all the material for a given skill.
- */
-const MaterialList: React.FC<MaterialSectionProps> = ({
+//TODO: Add documentation
+const MaterialList: React.FC<MaterialTabsProps> = ({
   materialKeys,
   defaultTab,
+  isCollapsible = true,
+  sectionName,
 }) => {
-  const [selectedTab, setSelectedTab] = useState(defaultTab || "");
-
-  if (!materialKeys || materialKeys.length === 0) {
-    return null;
-  }
-
-  const sections: MaterialSectionInterface[] = [
-    {
-      // Projects
-      name: MaterialType.Projects,
-      materials: projectKeys,
-      materialHashmap: projectDatabase,
-      basePath: PROJECTS_PAGE.path,
-      ListComponent: ProjectsList,
-    },
-    {
-      // University Modules
-      name: MaterialType.Modules,
-      materials: moduleKeys,
-      materialHashmap: moduleDatabase,
-      ListComponent: ModuleList,
-    },
-    {
-      name: MaterialType.WorkExperiences,
-      materials: roleKeys,
-      materialHashmap: rolesDatabase,
-      ListComponent: WorkList,
-      basePath: EXPERIENCE_PAGE.path,
-    },
-    {
-      // Certificates
-      name: MaterialType.Certificates,
-      materials: certificateKeys,
-      materialHashmap: certificateDatabase,
-      basePath: CERTIFICATES_PAGE.path,
-      ListComponent: CertificatesList,
-    },
-    {
-      // Blogs
-      name: MaterialType.Blogs,
-      materials: blogKeys,
-      materialHashmap: blogDatabase,
-      basePath: BLOG_PAGE.path,
-      ListComponent: BlogsList,
-    },
-  ];
-
-  // Filter out sections with no materials
-  const nonEmptySections: MaterialSectionInterface[] = sections.filter(
-    ({ materials, materialHashmap, name }) => {
-      const groupedMaterials: MaterialGroupInterface[] =
-        groupMaterialsByMaterialType(materialKeys, materialHashmap, name);
-      return (
-        groupedMaterials[0] && groupedMaterials[0].materialsKeys.length > 0
-      );
-    }
-  );
-
-  // Set default tab if none is selected
-  if (!selectedTab && nonEmptySections.length > 0) {
-    setSelectedTab(nonEmptySections[0].name);
-  }
-
-  return (
-    <Tabs
-      defaultValue={selectedTab}
-      className="w-full items-center md:items-start justify-center"
-      value={selectedTab}
-      onValueChange={setSelectedTab}
+  return isCollapsible ? (
+    <Accordion
+      type="single"
+      collapsible
+      className="
+        border-t-2 border-neutral-200 dark:border-neutral-700
+        p-3 mt-5"
     >
-      <TabsList
-        className="
-          w-full md:w-auto 
-          bg-transparent 
-          flex-col md:flex-row
-          -ml-4
-          "
-      >
-        {nonEmptySections.map(({ name }) => (
-          <TabsTrigger
-            key={name}
-            value={name}
+      <AccordionItem value="item-1">
+        <AccordionTrigger>
+          <p
             className="
-              text-2xl md:text-2xl font-bold
-              data-[state=inactive]:text-neutral-400 dark:data-[state=inactive]:text-neutral-600
-              data-[state=active]:shadow-none data-[state=active]:bg-transparent
-              transition-all duration-500
+              text-lg 
+              text-neutral-700 dark:text-neutral-300
               "
           >
-            {name}
-          </TabsTrigger>
-        ))}
-      </TabsList>
-
-      {nonEmptySections.map(
-        ({ name, materialHashmap, ListComponent, basePath }) => {
-          const groupedMaterials = groupMaterialsByMaterialType(
-            materialKeys,
-            materialHashmap,
-            name
-          );
-          return (
-            <TabsContent key={name} value={name}>
-              <div className="mt-4 text-center md:text-left">
-                <ListComponent groupedMaterial={groupedMaterials} />
-                {basePath && (
-                  <div className="flex justify-center mt-10">
-                    <Link href={basePath}>
-                      <Button variant="outline">{`View All ${name}`}</Button>
-                    </Link>
-                  </div>
-                )}
-              </div>
-            </TabsContent>
-          );
-        }
-      )}
-    </Tabs>
+            {`List of material directly related to ${sectionName}`}
+          </p>
+        </AccordionTrigger>
+        <AccordionContent>
+          <MaterialTab materialKeys={materialKeys} defaultTab={defaultTab} />
+        </AccordionContent>
+      </AccordionItem>
+    </Accordion>
+  ) : (
+    <MaterialTab materialKeys={materialKeys} defaultTab={defaultTab} />
   );
 };
